pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphagentservice"
version = "0.1.0"
description = "FastAPI service managed with uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dynaconf>=3.2.12",
    "fastapi>=0.135.1",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.6.6",
    "langchain-openai>=1.0.0",
    "openai>=1.68.2",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
graphagentservice = "graphagentservice.__main__:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/graphagentservice"]
